+++ src/main/java/com/nhom10/model/Car.java
@@ -16,8 +16,8 @@ public class Car {
     private String manufactured; // hang sx
     private String model;
     private String yearManufactured;
-    private int seats; // so ghe
-    private int longevity; // so nam sd
+    private Long seats; // so ghe
+    private Long longevity; // so nam sd
     private Date lastMaintenance; // ngay bao duong cuoi cung
 
     public Car() {
@@ -71,19 +71,19 @@ public class Car {
         this.yearManufactured = yearManufactured;
     }
 
-    public int getSeats() {
+    public Long getSeats() {
         return seats;
     }
 
-    public void setSeats(int seats) {
+    public void setSeats(Long seats) {
         this.seats = seats;
     }
 
-    public int getLongevity() {
+    public Long getLongevity() {
         return longevity;
     }
 
-    public void setLongevity(int longevity) {
+    public void setLongevity(Long longevity) {
         this.longevity = longevity;
     }
